fix(tree-console): fix tree node info panel build imports
