+++ app/src/router/pages/tree/console/TreeConsoleIntegrationLoader.tsx
@@ -1,0 +1,15 @@
+import { memo, lazy } from 'react';
+import type { TreeConsoleIntegrationProps } from '~/router/pages/tree/console/TreeConsoleIntegration';
+
+const LazyTreeConsoleIntegration = lazy(async () => {
+  const mod = await import('~/router/pages/tree/console/TreeConsoleIntegration');
+  return { default: mod.TreeConsoleIntegration };
+});
+
+export const MemoizedTreeConsoleIntegration = memo<TreeConsoleIntegrationProps>(
+  (props) => <LazyTreeConsoleIntegration {...props} />,
+  (prev, next) =>
+    prev.treeId === next.treeId &&
+    prev.pageNodeId === next.pageNodeId &&
+    (prev.pageTreeNode?.id ?? null) === (next.pageTreeNode?.id ?? null)
+);
