+++ app/src/router/pages/tree/console/hooks/useIndexedDbReset.ts
@@ -2,7 +2,7 @@ import { notify } from '@hierarchidb/components';
 import { useCallback } from 'react';
 import { useTranslation } from 'react-i18next';
 import { clearAppIndexedDBsViaPlugins } from '~/plugin-runtime/clearIndexedDb';
-import { logIntegrationWarning } from './treeConsoleIntegrationUtils.js';
+import { logIntegrationWarning } from '../treeConsoleIntegrationUtils.js';
 
 type IndexedDbResetArgs = {
   developerModeEnabled: boolean;
