+++ app/src/router/pages/tree/console/hooks/useTreeConsoleSpeedDial.ts
