+++ packages/ui/dynamic-speed-dial/src/index.ts
@@ -1,0 +1,11 @@
+export { DynamicSpeedDial } from './DynamicSpeedDial.js';
+export type { DynamicSpeedDialProps } from './DynamicSpeedDial.js';
+export { useDynamicSpeedDial } from './useDynamicSpeedDial.js';
+export type {
+  DynamicSpeedDialMenuItem,
+  DynamicSpeedDialNode,
+  DynamicSpeedDialTranslator,
+  DynamicSpeedDialIconResolver,
+  UseDynamicSpeedDialParams,
+  UseDynamicSpeedDialResult,
+} from './types.js';
