+++ packages/ui/dynamic-speed-dial/tsconfig.json
@@ -1,0 +1,11 @@
+{
+  "extends": "../../../tsconfig.package.local-alias.base.json",
+  "compilerOptions": {
+    "jsx": "react-jsx",
+    "allowImportingTsExtensions": true,
+    "allowArbitraryExtensions": true
+  },
+  "include": [
+    "src/**/*"
+  ]
+}
