+++ packages/ui/dynamic-speed-dial/tsconfig.typecheck.json
@@ -1,0 +1,3 @@
+{
+  "extends": "./tsconfig.json"
+}
